Refactor: rename triku.py to avoid circular import error.
